backend/requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
pymongo==3.11.4
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.2.4
uritemplate==4.1.1
